// web/src/api/generated/models/grading-stats-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ErrorCategoryCount } from './error-category-count';
// May contain unused imports in some cases
// @ts-ignore
import type { ScoreTrendItem } from './score-trend-item';

/**
 * 批改历史统计响应
 */
export interface GradingStatsResponse {
    /**
     * 总批改次数
     */
    'totalSubmissions'?: number;
    /**
     * 平均得分率（0.0~1.0）
     */
    'avgScoreRate'?: number;
    /**
     * 最近10次得分趋势 [{submissionId, score, maxScore, subject, createTime}]
     */
    'scoreTrend'?: Array<ScoreTrendItem>;
    /**
     * 学科得分分布 {subject: avgScoreRate}
     */
    'subjectScoreRates'?: { [key: string]: number; };
    /**
     * 错因分布（Top 错误类型及次数）
     */
    'errorDistribution'?: Array<ErrorCategoryCount>;
}