// web/src/api/generated/models/subject-profile-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { KnowledgeProfileResponse } from './knowledge-profile-response';

/**
 * 学科知识画像汇总
 */
export interface SubjectProfileSummary {
    /**
     * 学科
     */
    'subject'?: string;
    /**
     * 学科名称
     */
    'subjectName'?: string;
    /**
     * 平均掌握度
     */
    'avgMasteryLevel'?: number;
    /**
     * 总知识点数
     */
    'totalPoints'?: number;
    /**
     * 薄弱知识点数
     */
    'weakPointCount'?: number;
    /**
     * 优势知识点数（掌握度>=0.8）
     */
    'strongPointCount'?: number;
    /**
     * 薄弱知识点列表
     */
    'weakPoints'?: Array<KnowledgeProfileResponse>;
    /**
     * 优势知识点列表
     */
    'strongPoints'?: Array<KnowledgeProfileResponse>;
}