// web/src/api/generated/models/score-trend-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 得分趋势项
 */
export interface ScoreTrendItem {
    'submissionId'?: string;
    'score'?: number;
    'maxScore'?: number;
    'subject'?: string;
    'createTime'?: string;
}